requirements.txt
Django==5.2.10
openai==2.16.0
python-dotenv==1.2.1
pinecone[grpc]==8.0.0

PyJWT==2.10.1
cryptography==43.0.3

django-allauth>=65.14.0

openpyxl==3.1.5
xlrd==2.0.2
pypdf==6.6.2